feat: transaction detail view
