+++ LDKNodeMonday/View/Home/Payments/PaymentsListView.swift
@@ -17,6 +17,7 @@ struct PaymentsListView: View {
     @Binding var payments: [PaymentDetails]
     @Binding var displayBalanceType: DisplayBalanceType
     var price: Double
+    @State private var selectedPayment: PaymentDetails?
 
     var sections: [PaymentSection] {
         orderedStatuses.compactMap { status -> PaymentSection? in
@@ -66,6 +67,10 @@ struct PaymentsListView: View {
                         )
                         .padding(.vertical, 5)
                         .listRowSeparator(.hidden)
+                        .contentShape(Rectangle())
+                        .onTapGesture {
+                            self.selectedPayment = payment
+                        }
                     }
                 }
             }
@@ -79,6 +84,11 @@ struct PaymentsListView: View {
         }
         .listStyle(.plain)
         .padding(.horizontal)
+        .sheet(item: $selectedPayment) { paymentDetail in
+            PaymentDetailView(payment: paymentDetail)
+                .presentationDetents([.medium])
+                .presentationDragIndicator(.hidden)
+        }
     }
 }
 
@@ -101,8 +111,7 @@ struct PaymentItemView: View {
                     .frame(width: 16, height: 16)
                     .overlay(
                         Image(
-                            systemName: payment.kind
-                                == .onchain(txid: .init(), status: .unconfirmed)
+                            systemName: payment.isChainPayment
                                 ? "bitcoinsign" : "bolt"
                         )
                         .resizable()
@@ -141,119 +150,6 @@ struct PaymentItemView: View {
         .lineLimit(1)
         .minimumScaleFactor(0.75)
         .dynamicTypeSize(...DynamicTypeSize.accessibility2)  // Sets max dynamic size for all Text
-    }
-}
-
-extension PaymentDetails {
-    public var iconName: String {
-        switch self.status {
-        case .succeeded:
-            return self.direction == .inbound ? "arrow.down" : "arrow.up"
-        case .pending:
-            return "clock"
-        case .failed:
-            return "x.circle"
-        }
-    }
-
-    public var title: String {
-        switch self.status {
-        case .succeeded:
-            return self.direction == .inbound ? "Received" : "Sent"
-        case .pending:
-            return "Pending"
-        case .failed:
-            return "Failed"
-        }
-    }
-
-    public var formattedDate: String {
-        let calendar = Calendar.current
-        let now = Date.now
-        let date = Date(timeIntervalSince1970: TimeInterval(self.latestUpdateTimestamp))
-        let minutesSince = abs(date.timeIntervalSince(now)) / 60
-
-        switch minutesSince {
-        case ..<1:
-            return "Just now"
-
-        case ..<2:
-            return "1 minute ago"
-
-        case ..<60:
-            return "\(Int(minutesSince)) minutes ago"
-
-        default:
-            if calendar.isDate(date, inSameDayAs: now) {
-                return "Today at \(date.formatted(date: .omitted, time: .shortened))"
-            }
-
-            let dateFormat: Date.FormatStyle =
-                calendar.component(.year, from: date) == calendar.component(.year, from: now)
-                ? .dateTime.month(.abbreviated).day()
-                : .dateTime.month(.abbreviated).day().year()
-
-            return date.formatted(dateFormat)
-        }
-    }
-
-    public var amountColor: Color {
-        switch self.status {
-        case .succeeded:
-            return self.direction == .inbound ? .bitcoinGreen : .bitcoinNeutral8
-        case .pending:
-            return .bitcoinNeutral4
-        case .failed:
-            return .bitcoinNeutral4
-        }
-    }
-
-    public var secondaryAmountColor: Color {
-        switch self.status {
-        case .succeeded:
-            return .secondary
-        case .pending:
-            return .bitcoinNeutral4
-        case .failed:
-            return .bitcoinNeutral4
-        }
-    }
-
-    public func primaryAmount(displayBalanceType: DisplayBalanceType, price: Double) -> String {
-        let mSatsAmount = self.amountMsat ?? 0
-        let satsAmount = mSatsAmount.mSatsAsSats
-        let symbol = self.direction == .inbound ? "+ " : "- "
-
-        let formattedValue: String = {
-            switch displayBalanceType {
-            case .fiatSats, .fiatBtc:
-                return satsAmount.formattedSatsAsUSD(price: price)
-            case .btcFiat:
-                return satsAmount.formattedSatsAsBtc()
-            default:
-                return satsAmount.formatted(.number.notation(.automatic))
-            }
-        }()
-
-        return symbol + formattedValue
-    }
-
-    public func secondaryAmount(displayBalanceType: DisplayBalanceType, price: Double) -> String {
-        let mSatsAmount = self.amountMsat ?? 0
-        let satsAmount = mSatsAmount.mSatsAsSats
-
-        let formattedValue: String = {
-            switch displayBalanceType {
-            case .fiatSats:
-                return satsAmount.formatted(.number.notation(.automatic))
-            case .fiatBtc:
-                return satsAmount.formattedSatsAsBtc()
-            default:
-                return satsAmount.formattedSatsAsUSD(price: price)
-            }
-        }()
-
-        return formattedValue
     }
 }
